Worked on data acquisition using tghe usb port
For now a very simple interface is employed. I also wrote a python interface to the board.

Now I will have to write a XML-RPC server so that the computer can aquire the data. The esp32 will be connected to a raspberry pi and from there through a cabled network to the computer.