Started working on 8 anemometer board
The 16 anemometer board is working. But I think that 16 is too much. It would be better to use 8 anemometers per board.

The experience in building and using the multianem16 board and the esp32-daq-board showed some flaws in the design. Not related to the electronic circuit.

The following problems were observed:

 * Large current. For 32 anemometers, more than 3 A was observed. With a board with 8 anemometers, a current of 1 A per board would be sufficient.
 * More flexible usage.
 * Need to design a multi-temperature board.
 * Sharing the ground between different boards was clumsy. This problem was more noticeable in the daq board. In the new design, the flat cable from anemometer board to daq board will have 9 pins. One of them for the common ground.